requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asarPy==1.0.1
asyncio-dgram==3.0.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
Cython==3.2.4
dnspython==2.8.0
frozenlist==1.8.0
geoip2==5.2.0
html2image==2.0.7
idna==3.11
Mako==1.3.10.dev0
Markdown==3.10
MarkupSafe==3.0.3
maxminddb==3.1.1
mcstatus==13.0.0
meson==1.9.2
multidict==6.7.1
packaging==25.0
propcache==0.4.1
python-dotenv==1.2.2
requests==2.32.5
setuptools==80.9.0
urllib3==2.6.3
websocket-client==1.9.0
wheel==0.46.1
yarl==1.23.0
